new files

+++ script.sh
@@ -3,7 +3,10 @@
 function prepare_bastion() {
 sudo apt update
 sudo apt install ansible -y
-wget -O- https://apt.releases.hashicorp.com/gpg | sudo gpg --dearmor -o /usr/share/keyrings/hashicorp-archive-keyring.gpg
+if [ ! -f /usr/share/keyrings/hashicorp-archive-keyring.gpg ]; then
+    wget -O- https://apt.releases.hashicorp.com/gpg | sudo gpg --dearmor -o /usr/share/keyrings/hashicorp-archive-keyring.gpg
+fi
+
 echo "deb [signed-by=/usr/share/keyrings/hashicorp-archive-keyring.gpg] https://apt.releases.hashicorp.com $(lsb_release -cs) main" | sudo tee /etc/apt/sources.list.d/hashicorp.list
 sudo apt update && sudo apt install terraform
 }
@@ -11,9 +14,10 @@ sudo apt update && sudo apt install terraform
 function create_instances() {
     cd terraform
     terraform init
-    terraform apply -auto-approve
+    terraform apply --auto-approve
 }
 
 
 
 prepare_bastion
+create_instances
